refactor: 패키지 구조 변경 feat: 메인화면 테이블 리스트 구현
