+++ src/main/java/com/javabom/cafe/domain/menu/Category.java
@@ -1,0 +1,17 @@
+package com.javabom.cafe.domain.menu;
+
+public enum Category {
+    CAKE("케이크"),
+    BEVERAGE("음료");
+
+    private final String name;
+
+    Category(final String name) {
+        this.name = name;
+    }
+
+    @Override
+    public String toString() {
+        return "[" + name + "]";
+    }
+}
