+++ src/main/java/com/javabom/cafe/domain/menu/MenuRepository.java
@@ -1,0 +1,26 @@
+package com.javabom.cafe.domain.menu;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MenuRepository extends JpaRepository<Menu, Long> {
+//    private static final List<Menu> menus = new ArrayList<>();
+//
+//    static {
+//        menus.add(new Menu(1, "가나슈", Category.CAKE, 7000));
+//        menus.add(new Menu(2, "당근 케이크", Category.CAKE, 6500));
+//        menus.add(new Menu(3, "아이스 박스", Category.CAKE, 8000));
+//        menus.add(new Menu(4, "티라미수", Category.CAKE, 5500));
+//        menus.add(new Menu(5, "치즈 케이크", Category.CAKE, 7000));
+//        menus.add(new Menu(6, "얼그레이 케이크", Category.CAKE, 7000));
+//        menus.add(new Menu(21, "아메리카노", Category.BEVERAGE, 4500));
+//        menus.add(new Menu(22, "에스프레소", Category.BEVERAGE, 4000));
+//        menus.add(new Menu(23, "카푸치노", Category.BEVERAGE, 5000));
+//        menus.add(new Menu(24, "그린티 라떼", Category.BEVERAGE, 6000));
+//    }
+//
+//    public static List<Menu> menus() {
+//        return Collections.unmodifiableList(menus);
+//    }
+}
