+++ src/main/java/com/javabom/cafe/domain/table/Table.java
@@ -1,0 +1,34 @@
+package com.javabom.cafe.domain.table;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@NoArgsConstructor
+@Entity(name = "cafe_table")
+public class Table {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Getter
+    @Column(nullable = false)
+    private String tableName;
+
+    @Builder
+    public Table(final String tableName) {
+        this.tableName = tableName;
+    }
+
+    @Override
+    public String toString() {
+        return tableName;
+    }
+}
