+++ src/main/java/com/javabom/cafe/dto/table/TableInfoDto.java
@@ -1,0 +1,19 @@
+package com.javabom.cafe.dto.table;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class TableInfoDto {
+    private Long tableId;
+    private String tableName;
+    private boolean orderStatus;
+
+    @Builder
+    public TableInfoDto(final String tableName, final boolean orderStatus) {
+        this.tableName = tableName;
+        this.orderStatus = orderStatus;
+    }
+}
