+++ src/main/java/com/javabom/cafe/view/InputView.java
@@ -1,0 +1,12 @@
+package com.javabom.cafe.view;
+
+import java.util.Scanner;
+
+public class InputView {
+    private static final Scanner scanner = new Scanner(System.in);
+
+    public static int inputTableNumber() {
+        System.out.println("## 주문할 테이블을 선택하세요.");
+        return scanner.nextInt();
+    }
+}
