--- src/main/java/domain/menu/Category.java
@@ -1,17 +1,0 @@
-package domain.menu;
-
-public enum Category {
-    CAKE("케이크"),
-    BEVERAGE("음료");
-
-    private final String name;
-
-    Category(final String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String toString() {
-        return "[" + name + "]";
-    }
-}
