--- src/main/java/domain/menu/Menu.java
@@ -1,68 +1,0 @@
-package domain.menu;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import java.util.Objects;
-
-@NoArgsConstructor
-@Entity
-public class Menu {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private int number;
-
-    @Column(nullable = false)
-    @Getter
-    private String name;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    @Getter
-    private Category category;
-
-    @Column(nullable = false)
-    @Getter
-    private int price;
-
-    @Builder
-    public Menu(final int number, final String name, final Category category, final int price) {
-        this.number = number;
-        this.name = name;
-        this.category = category;
-        this.price = price;
-    }
-
-    @Override
-    public String toString() {
-        return category + " " + number + " - " + name + " : " + price + "원";
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Menu)) return false;
-        Menu menu = (Menu) o;
-        return number == menu.number &&
-                price == menu.price &&
-                Objects.equals(name, menu.name) &&
-                category == menu.category;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(number, name, category, price);
-    }
-}
