--- src/main/java/domain/order/OrderMenuRepository.java
@@ -1,12 +1,0 @@
-package domain.order;
-
-import domain.table.Table;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface OrderMenuRepository extends JpaRepository<OrderMenu, Long> {
-    List<OrderMenu> findByTable(final Table table);
-
-    boolean existsByTable(final Table table);
-}
