--- src/main/java/domain/table/Table.java
@@ -1,34 +1,0 @@
-package domain.table;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@NoArgsConstructor
-@Entity(name = "account")
-public class Table {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Getter
-    @Column(nullable = false)
-    private String tableName;
-
-    @Builder
-    public Table(final String tableName) {
-        this.tableName = tableName;
-    }
-
-    @Override
-    public String toString() {
-        return tableName;
-    }
-}
