--- src/main/java/dto/order/OrderAddDto.java
@@ -1,12 +1,0 @@
-package dto.order;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class OrderAddDto {
-    private Long tableId;
-    private Long menuId;
-    private int quantity;
-}
