--- src/main/java/view/InputView.java
@@ -1,12 +1,0 @@
-package view;
-
-import java.util.Scanner;
-
-public class InputView {
-    private static final Scanner scanner = new Scanner(System.in);
-
-    public static int inputTableNumber() {
-        System.out.println("## 주문할 테이블을 선택하세요.");
-        return scanner.nextInt();
-    }
-}
