--- src/main/java/view/OutputView.java
@@ -1,40 +1,0 @@
-package view;
-
-import domain.menu.Menu;
-import domain.table.Table;
-
-import java.util.List;
-
-public class OutputView {
-    private static final String TOP_LINE = "┌ ─ ┐";
-    private static final String TABLE_FORMAT = "| %s |";
-    private static final String BOTTOM_LINE = "└ ─ ┘";
-
-    public static void printTables(final List<Table> tables) {
-        System.out.println("## 테이블 목록");
-        final int size = tables.size();
-        printLine(TOP_LINE, size);
-        printTableNumbers(tables);
-        printLine(BOTTOM_LINE, size);
-    }
-
-    public static void printMenus(final List<Menu> menus) {
-        for (final Menu menu : menus) {
-            System.out.println(menu);
-        }
-    }
-
-    private static void printLine(final String line, final int count) {
-        for (int index = 0; index < count; index++) {
-            System.out.print(line);
-        }
-        System.out.println();
-    }
-
-    private static void printTableNumbers(final List<Table> tables) {
-        for (final Table table : tables) {
-            System.out.printf(TABLE_FORMAT, table);
-        }
-        System.out.println();
-    }
-}
