+++ src/main/resources/data-h2.sql
@@ -1,3 +1,3 @@
-insert into account(table_name) values ('1번 테이블');
-insert into account(table_name) values ('2번 테이블');
-insert into account(table_name) values ('3번 테이블');
+insert into cafe_table(table_name) values ('1번 테이블');
+insert into cafe_table(table_name) values ('2번 테이블');
+insert into cafe_table(table_name) values ('3번 테이블');
