+++ src/main/resources/static/js/api/index.js
@@ -33,8 +33,8 @@ const api = (() => {
   const requestWithJsonData = (uri, config) => fetch(uri, config).then(data => data.json())
 
   const table = {
-    getList() {
-      return request(`/tables`)
+    getAll() {
+        return request(`/tables`)
     },
     get(id) {
       return request(`/tables/${id}`)
